+++ android/src/main/java/com/cyanchill/missingcore/metadataretriever/models/TrackMetadataItem.kt
@@ -1,4 +1,4 @@
-package com.cyanchill.missingcore.metadataretriever.model
+package com.cyanchill.missingcore.metadataretriever.models
 
 data class TrackMetadataItem(
   override var bitrate: Int?,
